Move mergeability detection into the main pull loop. Add a munger that adds the "needs-rebase" label.
